Distributer Layout has been completed ,  some new network bugs they have corrected . Spinners Changed to ListViews
