+++ app/src/main/java/com/applicationforlife/jamesnikolaidis/thesis_farmers_helper/Network_Wifi/Network_Wifi_Class.java
@@ -83,7 +83,9 @@ public class Network_Wifi_Class {
         mWifiManager.setWifiEnabled(true);
     }
 
-
+    public void ChechMobileDataAndEnableIt(){
+        mApplicationActivity.enforceCallingOrSelfPermission(android.Manifest.permission.MODIFY_PHONE_STATE, null);
+    }
