+++ app/src/main/res/layout/distributer_details_layout.xml
@@ -1,0 +1,210 @@
+<?xml version="1.0" encoding="utf-8"?>
+<ScrollView android:layout_width="match_parent"
+    android:layout_height="match_parent"
+
+    xmlns:android="http://schemas.android.com/apk/res/android" >
+
+
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical" android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@drawable/toolbar_shape">
+
+
+    <TextView
+        android:layout_marginTop="5dp"
+        android:layout_width="wrap_content"
+        android:layout_height= "wrap_content"
+        android:layout_gravity="center"
+        android:textAlignment="center"
+        android:text=" Distributor Name"
+        android:textColor="@android:color/white"
+        android:textSize="24dp"
+        android:fontFamily="sans-serif-medium"
+
+
+        />
+    <TextView
+        android:layout_marginTop="12dp"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textSize="20dp"
+        android:gravity="center"
+        android:layout_gravity="center"
+        android:textAlignment="center"
+        android:background="@drawable/tip_shape"
+        android:fontFamily="sans-serif-medium"
+        android:textColor="@android:color/white"
+        android:text="Nothing Found..."
+        android:id="@+id/DistributerNameTextView"
+        />
+
+
+
+    <TextView
+        android:layout_marginTop="20dp"
+        android:layout_width="wrap_content"
+        android:layout_height= "wrap_content"
+        android:layout_gravity="center"
+        android:textAlignment="center"
+        android:text=" Distributor Email"
+        android:textColor="@android:color/white"
+        android:textSize="24dp"
+        android:fontFamily="sans-serif-medium"
+
+
+        />
+
+    <TextView
+        android:layout_marginTop="12dp"
+        android:layout_width="250dp"
+        android:layout_height="wrap_content"
+        android:textSize="20dp"
+        android:layout_gravity="center"
+        android:textAlignment="center"
+        android:gravity="center"
+        android:background="@drawable/tip_shape"
+        android:fontFamily="sans-serif-medium"
+        android:textColor="@android:color/white"
+        android:text="Nothing Found..."
+        android:id="@+id/DistributerEmailTextView"
+        />
+
+    <TextView
+        android:layout_marginTop="20dp"
+        android:layout_width="wrap_content"
+        android:layout_height= "wrap_content"
+        android:layout_gravity="center"
+        android:textAlignment="center"
+
+        android:text=" Distributor Address"
+        android:textColor="@android:color/white"
+        android:textSize="24dp"
+        android:fontFamily="sans-serif-medium"
+
+
+        />
+    <TextView
+        android:layout_marginTop="12dp"
+        android:layout_width="250dp"
+        android:layout_height="wrap_content"
+        android:textSize="20dp"
+        android:layout_gravity="center"
+        android:gravity="center"
+        android:textAlignment="center"
+        android:background="@drawable/tip_shape"
+        android:fontFamily="sans-serif-medium"
+        android:textColor="@android:color/white"
+        android:id="@+id/DistributerAddressTextView"
+        android:text="Nothing Found..."
+        />
+
+    <TextView
+        android:layout_marginTop="20dp"
+        android:layout_width="wrap_content"
+        android:layout_height= "wrap_content"
+        android:layout_gravity="center"
+        android:textAlignment="center"
+        android:text=" Distributor Country"
+        android:textColor="@android:color/white"
+        android:textSize="24dp"
+        android:fontFamily="sans-serif-medium"
+
+
+        />
+    <TextView
+        android:layout_marginTop="12dp"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textSize="20dp"
+        android:layout_gravity="center"
+        android:textAlignment="center"
+        android:background="@drawable/tip_shape"
+        android:fontFamily="sans-serif-medium"
+        android:textColor="@android:color/white"
+        android:text="Nothing Found..."
+        android:id="@+id/DistributerCountryTextView"
+        android:gravity="center"
+        />
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal"
+        android:layout_marginTop="20dp">
+        <TextView
+            android:layout_marginLeft="5dp"
+            android:layout_width="wrap_content"
+            android:layout_height= "wrap_content"
+            android:layout_gravity="center"
+            android:textAlignment="center"
+            android:text=" Distributor State"
+            android:textColor="@android:color/white"
+            android:textSize="20dp"
+            android:fontFamily="sans-serif-medium"
+
+
+            />
+        <TextView
+            android:layout_marginLeft="20dp"
+            android:layout_width="wrap_content"
+            android:layout_height= "wrap_content"
+            android:layout_gravity="center"
+            android:textAlignment="center"
+            android:text=" Distributor Postal"
+            android:textColor="@android:color/white"
+            android:textSize="20dp"
+            android:fontFamily="sans-serif-medium"
+
+
+            />
+    </LinearLayout>
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal"
+        android:layout_marginTop="20dp">
+        <TextView
+
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textSize="20dp"
+            android:layout_gravity="center"
+            android:textAlignment="center"
+            android:background="@drawable/tip_shape"
+            android:fontFamily="sans-serif-medium"
+            android:textColor="@android:color/white"
+            android:text="Nothing Found..."
+            android:id="@+id/DistributerStateTextView"
+            android:gravity="center"
+            />
+
+        <TextView
+            android:layout_marginLeft="10dp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textSize="20dp"
+            android:layout_gravity="center"
+            android:textAlignment="center"
+            android:background="@drawable/tip_shape"
+            android:fontFamily="sans-serif-medium"
+            android:textColor="@android:color/white"
+            android:text="Nothing Found..."
+            android:id="@+id/DistributerPostalTextView"
+            android:gravity="center"
+            />
+
+    </LinearLayout>
+
+
+
+    <Button
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:id="@+id/DistClose"
+        android:background="@drawable/chat_buttons_shape"
+            android:text="Close"/>
+</LinearLayout>
+
+</ScrollView>
